feat: Show cards on home page based on login button selected
