+++ src/components/App/App.jsx
@@ -1,8 +1,7 @@
 import './App.css';
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { Routes, Route, Link } from 'react-router-dom';
-import { useApolloClient, gql } from '@apollo/client';
 import Home from '../Home/Home.jsx';
 import Navigation from '../Navigation/Navigation.jsx';
 import EncounterBuilder from '../EncounterBuilder/EncounterBuilder.jsx';
@@ -12,40 +11,7 @@ import ErrorPage from '../ErrorPage/ErrorPage.jsx';
 
 export default function App() {
   const [encounters, setEncounters] = useState([]);
-  const client = useApolloClient();
-
-  useEffect(() => {
-    const getEncounters = async () => {
-      try {
-        const { data } = await client.query({
-          query: gql`
-            query getEncounters($userName: String!) {
-              encounters(userName: $userName) {
-                  id
-                  userName
-                  encounterName
-                  partySize
-                  partyLevel
-                  summary
-                  description
-                  treasure
-                  encounterMonsters {
-                      monsterName
-                  }
-              }
-          }
-    `,
-          variables: {
-            "userName": "demo-many-encounters"
-          },
-        });
-        setEncounters(data.encounters);
-      } catch (error) {
-        console.error('Error fetching encounters: ', error);
-      }
-    };
-    getEncounters();
-  }, [client]);
+  const [userName, setUserName] = useState(null);
 
   return (
     <div className="App">
@@ -56,8 +22,8 @@ export default function App() {
         <Navigation />
       </header>
       <Routes>
-        <Route path="/" element={<Home encounters={encounters} />} />
-        <Route path="/login" element={<Login />} />
+        <Route path="/" element={<Home userName={userName} />} />
+        <Route path="/login" element={<Login setUserName={setUserName} />} />
         <Route path="/details" element={<EncounterDetails encounters={encounters} />} />
         <Route path="/encounterbuilder" element={<EncounterBuilder />} />
         <Route path="*" element={<ErrorPage />} />
